requirements.txt
earthengine-api
google-cloud-bigquery[pandas, geopandas, bqstorage]
plotly
matplotlib
kaleido